requirements.txt
aiohttp==1.3.3
python-socketio==1.7.1
